// packages/frontend/src/generated/types/WarehouseResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

/**
 * @description Response DTO for the `warehouse` entity.
 */
export interface WarehouseResponse {
  /**
   * @type string, uuid
   */
  baseId: string
  /**
   * @type string
   */
  commonName: string
  /**
   * @type string
   */
  createdAt: string
  /**
   * @type string, uuid
   */
  createdBy: string
  /**
   * @type string,null
   */
  deletedAt?: string | null
  /**
   * @type string,null, uuid
   */
  deletedBy?: string | null
  /**
   * @type string, uuid
   */
  id: string
  /**
   * @type string,null
   */
  longName?: string | null
  /**
   * @type string, uuid
   */
  originDbId: string
  /**
   * @type string
   */
  updatedAt: string
  /**
   * @type string, uuid
   */
  updatedBy: string
}